+++ Sources/SwiftSDKCommand/SwiftSDKCommand.swift
@@ -21,6 +21,7 @@ package struct SwiftSDKCommand: AsyncParsableCommand {
         version: SwiftVersion.current.completeDisplayString,
         subcommands: [
             ConfigureSwiftSDK.self,
+            DeprecatedSwiftSDKConfigurationCommand.self,
             InstallSwiftSDK.self,
             ListSwiftSDKs.self,
             RemoveSwiftSDK.self,